refactor(Android, Stack): Split large classes, reorganize package structure
Related commit from RNScreens: https://github.com/software-mansion/react-native-screens/commit/d6e39013fb8e3fba34c4e8dcdb0bfb2b5ce70238

Refactors the Android header implementation: splits large files/classes
and reorganizes the package structure to match RNS.

- app-bar files move to header/appbar (StackHeaderAppBarLayout,
  StackHeaderAppBarLayoutBehavior, StackHeaderScrollingViewBehavior)
- menu config/model types move to toolbar/model; the field-update,
  element-options and update-queue types move to toolbar/update, with
  the queue file split into five (queue, raw update, resolved update,
  icon resolver, queue delegate)
- StackHeaderToolbarMenuApplicator extracted from StackHeaderApplicator
  (menu build, mappings/metadata, element updates, icon tint)
- StackHeaderFrameSynchronizer (header frame + subview origin sync) and
  StackHeaderToolbarMenuSelectionController (id map + group metadata +
  checked-state changes) extracted from StackHeaderCoordinatorLayout
- comments added to header-related classes

Lynx divergences carry through the moves unchanged: the subviewView
accessor, the content-behavior no-op divergence note, back-press via
the layout's activity context, and UI-method wording in the queue docs.

Verified on emulator: the batch commands example behaves identically
after the restructuring (one coalesced fruits event for Select All, a
cross-group batch emits only for the group whose selection changed).
