+++ android/src/main/java/com/lynxscreens/screens/header/StackHeaderCoordinatorLayout.kt
@@ -11,16 +11,23 @@ import androidx.coordinatorlayout.widget.CoordinatorLayout
 import com.google.android.material.R
 import com.google.android.material.appbar.AppBarLayout
 import com.google.android.material.appbar.MaterialToolbar
+import com.lynxscreens.screens.header.appbar.StackHeaderAppBarLayout
+import com.lynxscreens.screens.header.appbar.StackHeaderScrollingViewBehavior
 import com.lynxscreens.screens.header.config.OnHeaderConfigurationAttachListener
 import com.lynxscreens.screens.header.config.StackHeaderConfigurationObserver
 import com.lynxscreens.screens.header.config.StackHeaderConfigurationProviding
 import com.lynxscreens.screens.header.config.StackHeaderDelegate
 import com.lynxscreens.screens.header.config.StackHeaderInvalidationFlags
-import com.lynxscreens.screens.header.subview.StackHeaderSubviewProviding
-import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuGroupMetadata
-import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuElementUpdate
+import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuApplicator
+import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuSelectionController
+import com.lynxscreens.screens.header.toolbar.update.StackHeaderToolbarMenuElementUpdate
 import com.lynxscreens.screens.screen.StackScreenComponent
 
+/**
+ * Root CoordinatorLayout for a screen's header: hosts the app bar and the
+ * content wrapper, wires header-config attach/observe, routes config
+ * invalidations to the applicators, and owns header lifecycle/teardown.
+ */
 @SuppressLint("ViewConstructor")
 internal class StackHeaderCoordinatorLayout(
     context: Context,
@@ -79,15 +86,15 @@ internal class StackHeaderCoordinatorLayout(
                 // then emit a single coalesced event per affected group.
                 val affectedGroups = LinkedHashSet<String>()
                 for (update in updates) {
-                    applicator.updateToolbarMenuElement(
+                    StackHeaderToolbarMenuApplicator.updateToolbarMenuElement(
                         toolbar,
-                        toolbarMenuForwardIdMap,
+                        selectionController.forwardIdMap,
                         update.id,
                         update.options,
                     )
                     val checked = update.options.checked
                     if (checked != null) {
-                        applyGroupItemStateChange(toolbar, update.id, checked)?.let(affectedGroups::add)
+                        selectionController.applyGroupItemStateChange(toolbar, update.id, checked)?.let(affectedGroups::add)
                     }
                 }
                 affectedGroups.forEach { groupId -> emitGroupSelection(toolbar, groupId) }
@@ -122,49 +129,7 @@ internal class StackHeaderCoordinatorLayout(
         val delegate = currentDelegate ?: return
         val provider = currentProvider ?: return
         val appBar = appBarLayout ?: return
-
-        // When config is transparent, the StackScreen is static so we need to offset the header
-        // config by the offset of the AppBarLayout (which is 0 or is negative). When config is
-        // opaque, the Screen always moves with the config, that's why we need to offset the
-        // header config by the negative value of AppBarLayout's height.
-        val configOffset = if (provider.transparent) appBar.top else appBar.top - appBar.bottom
-
-        delegate.onHeaderFrameChanged(
-            appBar.width,
-            appBar.height,
-            configOffset,
-        )
-
-        updateSubviewOffsets(appBar, provider)
-    }
-
-    private fun updateSubviewOffsets(
-        appBar: StackHeaderAppBarLayout,
-        config: StackHeaderConfigurationProviding,
-    ) {
-        config.leadingSubview?.let { updateSubviewOffset(it, appBar) }
-        config.centerSubview?.let { updateSubviewOffset(it, appBar) }
-        config.trailingSubview?.let { updateSubviewOffset(it, appBar) }
-        config.backgroundSubview?.let { updateSubviewOffset(it, appBar) }
-    }
-
-    private fun updateSubviewOffset(
-        subview: StackHeaderSubviewProviding,
-        appBar: StackHeaderAppBarLayout,
-    ) {
-        val view = subview.subviewView
-        if (view.width == 0 && view.height == 0) return
-
-        val appBarPos = IntArray(2)
-        val subviewPos = IntArray(2)
-        appBar.getLocationInWindow(appBarPos)
-        view.getLocationInWindow(subviewPos)
-
-        currentDelegate?.onSubviewOriginChanged(
-            subview.type,
-            x = subviewPos[0] - appBarPos[0],
-            y = subviewPos[1] - appBarPos[1],
-        )
+        StackHeaderFrameSynchronizer.sync(appBar, provider, delegate)
     }
 
     // endregion
@@ -179,11 +144,9 @@ internal class StackHeaderCoordinatorLayout(
 
     private val applicator = StackHeaderApplicator(wrappedContext)
 
+    private val selectionController = StackHeaderToolbarMenuSelectionController()
+
     private var appBarLayout: StackHeaderAppBarLayout? = null
-
-    private var toolbarMenuForwardIdMap = emptyMap<String, Int>()
-
-    private var toolbarMenuGroupMetadata = StackHeaderToolbarMenuGroupMetadata.EMPTY
 
     private val onNavigationIconClick: () -> Unit = {
         // The fragment constructs this layout with its (activity) context.
@@ -232,24 +195,23 @@ internal class StackHeaderCoordinatorLayout(
 
             if (provider.invalidationFlags.containsAny(StackHeaderInvalidationFlags.TOOLBAR_MENU)) {
                 val (forwardIdMap, reverseIdMap) =
-                    applicator.generateToolbarMenuItemMappings(
+                    StackHeaderToolbarMenuApplicator.generateToolbarMenuItemMappings(
                         provider.toolbarMenu,
                     )
                 val forwardGroupIdMap =
-                    applicator.generateToolbarMenuGroupMappings(
+                    StackHeaderToolbarMenuApplicator.generateToolbarMenuGroupMappings(
                         provider.toolbarMenu,
                     )
                 val groupMetadata =
-                    applicator.computeGroupMetadata(
+                    StackHeaderToolbarMenuApplicator.computeGroupMetadata(
                         provider.toolbarMenu,
                     )
 
-                applicator.validateRadioInitialSelection(provider.toolbarMenu)
+                StackHeaderToolbarMenuApplicator.validateRadioInitialSelection(provider.toolbarMenu)
 
-                toolbarMenuForwardIdMap = forwardIdMap
-                toolbarMenuGroupMetadata = groupMetadata
+                selectionController.setMenuMaps(forwardIdMap, groupMetadata)
 
-                applicator.rebuildToolbarMenu(
+                StackHeaderToolbarMenuApplicator.rebuildToolbarMenu(
                     appBar.toolbar,
                     provider.toolbarMenu,
                     forwardIdMap,
@@ -258,7 +220,7 @@ internal class StackHeaderCoordinatorLayout(
                     groupDividerEnabled = provider.toolbarMenuGroupDividerEnabled,
                     onItemClicked = { id, menuItem ->
                         if (menuItem.isCheckable) {
-                            applyGroupItemStateChange(appBar.toolbar, id)?.let { groupId ->
+                            selectionController.applyGroupItemStateChange(appBar.toolbar, id)?.let { groupId ->
                                 emitGroupSelection(appBar.toolbar, groupId)
                             }
                         } else {
@@ -278,62 +240,12 @@ internal class StackHeaderCoordinatorLayout(
 
     // region Group selection
 
-    /**
-     * Mutates the checked state of [itemId] within its group and returns the id of the group
-     * whose selection changed, or `null` when nothing changed (the item is not in a group, an
-     * invalid attempt to uncheck a single-selection item, or the item was already in the
-     * target state). Does not emit — callers decide when to emit so that batched updates can
-     * coalesce into one event per group.
-     */
-    private fun applyGroupItemStateChange(
-        toolbar: MaterialToolbar,
-        itemId: String,
-        explicitCheckedValue: Boolean? = null,
-    ): String? {
-        val groupId = toolbarMenuGroupMetadata.itemGroupMap[itemId] ?: return null
-        val singleSelection = toolbarMenuGroupMetadata.groupSingleSelection[groupId] ?: return null
-        val intId = toolbarMenuForwardIdMap[itemId] ?: return null
-        val menuItem = toolbar.menu.findItem(intId) ?: return null
-
-        if (singleSelection && explicitCheckedValue == false) {
-            Log.w(
-                TAG,
-                "[RNScreens] Cannot uncheck item '$itemId' in single-selection group '$groupId'. " +
-                    "Check a different item instead.",
-            )
-            return null
-        }
-
-        val newChecked =
-            if (singleSelection) {
-                true
-            } else {
-                explicitCheckedValue ?: !menuItem.isChecked
-            }
-        if (menuItem.isChecked == newChecked) return null
-        menuItem.isChecked = newChecked
-
-        return groupId
-    }
-
     private fun emitGroupSelection(
         toolbar: MaterialToolbar,
         groupId: String,
     ) {
-        currentDelegate?.onGroupSelectionChanged(groupId, collectSelectedIds(toolbar, groupId))
+        currentDelegate?.onGroupSelectionChanged(groupId, selectionController.collectSelectedIds(toolbar, groupId))
     }
-
-    private fun collectSelectedIds(
-        toolbar: MaterialToolbar,
-        groupId: String,
-    ): List<String> =
-        toolbarMenuGroupMetadata
-            .groupMemberItems[groupId]
-            .orEmpty()
-            .filter { memberId ->
-                val intId = toolbarMenuForwardIdMap[memberId] ?: return@filter false
-                toolbar.menu.findItem(intId)?.isChecked == true
-            }
 
     // endregion
 
@@ -345,8 +257,7 @@ internal class StackHeaderCoordinatorLayout(
             removeView(it)
         }
         appBarLayout = null
-        toolbarMenuForwardIdMap = emptyMap()
-        toolbarMenuGroupMetadata = StackHeaderToolbarMenuGroupMetadata.EMPTY
+        selectionController.clear()
     }
 
     private fun removeHeader() {
