+++ android/src/main/java/com/lynxscreens/screens/header/config/StackHeaderConfigComponent.kt
@@ -18,13 +18,13 @@ import com.lynxscreens.screens.common.ShadowStateProxy
 import com.lynxscreens.screens.header.subview.OnStackHeaderSubviewChangeListener
 import com.lynxscreens.screens.header.subview.StackHeaderSubviewComponent
 import com.lynxscreens.screens.header.subview.StackHeaderSubviewType
-import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuConfig
-import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuElementRawUpdate
-import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuIconResolver
-import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuItemIconSource
+import com.lynxscreens.screens.header.toolbar.model.StackHeaderToolbarMenuConfig
+import com.lynxscreens.screens.header.toolbar.update.StackHeaderToolbarMenuElementRawUpdate
+import com.lynxscreens.screens.header.toolbar.update.StackHeaderToolbarMenuIconResolver
+import com.lynxscreens.screens.header.toolbar.model.StackHeaderToolbarMenuItemIconSource
 import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuMapper
-import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuUpdateQueue
-import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarFieldUpdate
+import com.lynxscreens.screens.header.toolbar.update.StackHeaderToolbarMenuUpdateQueue
+import com.lynxscreens.screens.header.toolbar.update.StackHeaderToolbarFieldUpdate
 import com.lynxscreens.screens.helpers.IconResolution
 import com.lynxscreens.screens.helpers.PropIconResolver
 import com.lynxscreens.screens.helpers.resolveImage
