+++ android/src/main/java/com/lynxscreens/screens/header/config/StackHeaderConfigurationObserver.kt
@@ -1,6 +1,6 @@
 package com.lynxscreens.screens.header.config
 
-import com.lynxscreens.screens.header.toolbar.StackHeaderToolbarMenuElementUpdate
+import com.lynxscreens.screens.header.toolbar.update.StackHeaderToolbarMenuElementUpdate
 
 internal interface StackHeaderConfigurationObserver {
     fun onConfigChanged(config: StackHeaderConfigurationProviding)
