+++ android/src/main/java/com/lynxscreens/screens/header/toolbar/model/StackHeaderToolbarMenuGroupMetadata.kt
@@ -1,5 +1,9 @@
-package com.lynxscreens.screens.header.toolbar
+package com.lynxscreens.screens.header.toolbar.model
 
+/**
+ * Derived group structure for a built menu: item→group, per-group
+ * single-selection, and group membership.
+ */
 internal data class StackHeaderToolbarMenuGroupMetadata(
     val itemGroupMap: Map<String, String>,
     val groupSingleSelection: Map<String, Boolean>,
