+++ android/src/main/java/com/lynxscreens/screens/header/toolbar/model/StackHeaderToolbarMenuItemConfig.kt
@@ -1,4 +1,4 @@
-package com.lynxscreens.screens.header.toolbar
+package com.lynxscreens.screens.header.toolbar.model
 
 import android.graphics.drawable.Drawable
