+++ android/src/main/java/com/lynxscreens/screens/header/toolbar/model/StackHeaderToolbarMenuItemType.kt
@@ -1,4 +1,4 @@
-package com.lynxscreens.screens.header.toolbar
+package com.lynxscreens.screens.header.toolbar.model
 
 internal enum class StackHeaderToolbarMenuItemType {
     ACTION,
