+++ android/src/main/java/com/lynxscreens/screens/header/toolbar/update/StackHeaderToolbarMenuUpdateQueueDelegate.kt
@@ -1,0 +1,9 @@
+package com.lynxscreens.screens.header.toolbar.update
+
+/**
+ * Receives a fully-resolved batch of menu element updates, to be applied to the toolbar
+ * atomically.
+ */
+internal fun interface StackHeaderToolbarMenuUpdateQueueDelegate {
+    fun onUpdatesResolved(updates: List<StackHeaderToolbarMenuElementUpdate>)
+}
