+++ docs/rns-sync-plan.md
@@ -66,7 +66,7 @@ Lanes: `core` = existing lynx-screens code (Host/Screen/model), `hdr-and` / `hdr
 - [x] `693082b36` L hdr-ios — feat(Stack v5, iOS): Implement basic view commands for iOS header menu ([#4339](https://github.com/software-mansion/react-native-screens/commit/693082b36)) — ported 1:1 (setMenuItemOptions/setMenuOptions commands changing title/icon/toggleState, RNSStackHeaderMenuFinder + RNSStackHeaderMenuUpdateOptions, immutable menu-tree replacement in the menu coordinator, item keeps prop copies via titleProp/iconProp/menuProp, tracker gains setToggleState/toggleStateEquals and drops toggleStateChanged - selection events now always emit and re-selecting a radio early-returns, coordinator setToggleState resolves the single-selection root and emits the selection event); adaptations: commands are LYNX_UI_METHODs taking {menuElementId, options} params (no codegen array wrapper - extractOptionsDict reads the options key), RCTAssert/RCTLogWarn → NSAssert/LLogWarn, no icon asset resolution step on the JS side (uri passthrough); menu example gains command pickers; verified on simulator via a temporary on-mount effect + lldb (action re-titled 'Changed' with star.fill, submenu re-titled 'New Title' with heart.fill, toggleState:true sets UIAction state=On and delivers Selected "toggle-1-2" to JS)
 - [x] `9c62292d8` S core-and — chore(Android): bump Material to 1.14.0 ([#4348](https://github.com/software-mansion/react-native-screens/commit/9c62292d8)) — added to the queue 2026-08-20 (Material 1.14 is important in RNS and a prerequisite for the M3 Expressive App Bar theme); bumped 1.13.0 → 1.14.0 in android/build.gradle.kts; regression-checked on emulator per the RNS Stack v5 test plan (toolbar menu icons render, large collapsing header renders)
 - [x] `96a4107b1` S hdr-and — feat(Android, Stack v5): use M3 Expressive theme for App Bar ([#4365](https://github.com/software-mansion/react-native-screens/commit/96a4107b1)) — ported 1:1 (wrappedContext theme → Theme_Material3Expressive_DayNight_NoActionBar); verified on emulator (large header renders in the Expressive style - larger title/height than plain M3)
-- [ ] `d6e39013f` L core-and — refactor(Android, Stack v5): split large classes, reorganize package structure ([#4354](https://github.com/software-mansion/react-native-screens/commit/d6e39013f)) — align lynx-screens package layout with RNS
+- [x] `d6e39013f` L core-and — refactor(Android, Stack v5): split large classes, reorganize package structure ([#4354](https://github.com/software-mansion/react-native-screens/commit/d6e39013f)) — ported 1:1 (appbar files → header/appbar, menu config/model types → toolbar/model, field-update/options/queue → toolbar/update with the queue split into five files, StackHeaderToolbarMenuApplicator extracted from StackHeaderApplicator, StackHeaderFrameSynchronizer + StackHeaderToolbarMenuSelectionController extracted from StackHeaderCoordinatorLayout, class docs added); Lynx divergences preserved through the moves (subviewView accessor, content-behavior no-op divergence note, back-press via layout context, UI-method wording in queue docs); regression-verified on emulator (batch commands example behaves identically: one coalesced fruits event, cross-group batch emits only for the changed group)
 - [ ] `2828bb2a5` M hdr-and — feat(Android, Stack v5): expose `liftOnScroll` prop for `small` header, integrate with SVM ([#4368](https://github.com/software-mansion/react-native-screens/commit/2828bb2a5)) — depends on ScrollViewMarker, which lynx-screens doesn't have; decide scope on review
 - [ ] `cff4f4131` L hdr-ios — feat(Stack v5, iOS): Add titleMenu ([#4349](https://github.com/software-mansion/react-native-screens/commit/cff4f4131))
 - [ ] `406318890` S example — refactor(examples, stack v5): erase options as required from StackRouteConfig ([#4419](https://github.com/software-mansion/react-native-screens/commit/406318890))
